% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.create_results.R
\name{add_allele_features}
\alias{add_allele_features}
\title{Add Allele Balance Features to DataFrame}
\usage{
add_allele_features(df, b = 9)
}
\arguments{
\item{df}{input data frame must contain the character columns AD_MUT and AD_WT}

\item{b}{integer, defining the threshold multiple of "strong advantage" (default =9)}
}
\value{
returns the data frame with 8 new columns added, and the column name format is:
\itemize{
  \item \code{MUT_strong_alt}, \code{MUT_strong_ref}, \code{MUT_weak_alt}, \code{MUT_weak_ref}
  \item \code{WT_strong_alt}, \code{WT_strong_ref}, \code{WT_weak_alt}, \code{WT_weak_ref}
}
}
\description{
Add allelic balance feature columns to data frames containing AD_MUT and AD_WT columns. Four feature columns will be generated for each sample
The result data frame has added 8 columns (4 columns each for MUT_* and WT_*).
}
\examples{
# 创建示例数据框
test_df <- data.frame(
  ID = c("snp1", "snp2"),
  AD_MUT = c("10,90", "50,5"),
  AD_WT = c("15,10", "10,10")
)

# 添加特征列
add_allele_features(test_df)
}
