% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.create_results.R
\name{create_results_from_vcf}
\alias{create_results_from_vcf}
\title{Title}
\usage{
create_results_from_vcf(
  outdir,
  sample_mut,
  sample_wt,
  depthFT = "False",
  mdepth = 0,
  GT_mut = "aa",
  mindp = 4,
  snp_vcf
)
}
\arguments{
\item{outdir}{outdir}

\item{sample_mut}{mutant sample}

\item{sample_wt}{wild type sample}

\item{depthFT}{Whether to enable depth filtering (default "False"), use "True" to enable}

\item{mdepth}{specified sequencing depth threshold (default 0, specify non-zero value), effective when depthFT="True"}

\item{GT_mut}{mutant genotype}

\item{mindp}{Minimum depth threshold (default 4) for filtering variant sites}

\item{snp_vcf}{Path to SNP VCF file}
}
\value{
No direct return value. Result files will be written to `outdir` directory, including:
}
\description{
Title
}
\examples{
create_results("/path/to/outdir", "MUT_sample", "WT_sample", depthFT = "False", GT_mut = "aa", mindp = 4)
}
