% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.create_results.R
\name{extract_sample_features}
\alias{extract_sample_features}
\title{Extract Allele Balance Features from AD String}
\usage{
extract_sample_features(ad_str, b = 9)
}
\arguments{
\item{ad_str}{string, allele depth in AD format ("ref_depth,alt_depth")}

\item{b}{integer, defining the threshold multiple of "strong advantage" (default =9)}
}
\value{
returns an integer vector of length 4, which contains the following features:
\itemize{
\item \code{strong_alt}: Depth of variant allele ≥ b * depth of reference allele
\item \code{strong_ref}: Reference allele depth ≥ b * variant allele depth
\item \code{weak_alt}: The depth of the variant allele is greater than that of the reference allele and less than b times the reference depth
\item \code{weak_ref}: The depth of the reference allele is greater than that of the variant allele and less than b times the variant depth
}
}
\description{
Extract four allelic balance features from the AD format string ("ref_depth,alt_depth") :
Strong variation advantage, strong reference advantage, weak variation advantage, weak reference advantage.
}
\examples{
extract_sample_features("10,90")  # c(1, 0, 0, 0)
extract_sample_features("50,5")   # c(0, 1, 0, 0)
extract_sample_features("10,15")  # c(0, 0, 1, 0)
extract_sample_features("15,10")  # c(0, 0, 0, 1)
extract_sample_features("10,10")  # c(0, 0, 0, 0)
}
