% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aa03.getgraphsfromxlsx.R
\name{get_graphs}
\alias{get_graphs}
\title{Title plot}
\usage{
get_graphs(file, outdir, sample_mut)
}
\arguments{
\item{file}{input xlsx file(from get_table function)}

\item{outdir}{output directory}

\item{sample_mut}{mutant sample name}
}
\value{
null
}
\description{
plot
}
\examples{
get_graphs("B3304.snp.VQSR.snpEff.xlsx")
}
