% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_software_path.R
\name{get_software_paths}
\alias{get_software_paths}
\title{Get currently configured software paths}
\usage{
get_software_paths()
}
\value{
A list of software paths
}
\description{
Get currently configured software paths
}
