% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aa02.getxlsxfromtxt.R
\name{get_table}
\alias{get_table}
\title{Title get_table}
\usage{
get_table(
  file,
  chrlist,
  mindp = 4,
  minSNPindexmut = 0.3,
  maxSNPindexmut = 0.7,
  outdir,
  M
)
}
\arguments{
\item{file}{txt file path(from get_txt function)}

\item{chrlist}{chromosome list}

\item{mindp}{Minimum depth for filtering (default: 4)}

\item{minSNPindexmut}{minimum SNPindex-M (default:0.3)}

\item{maxSNPindexmut}{maximum SNPindex-M (default:0.7)}

\item{outdir}{output directory}

\item{M}{mutant sample name}
}
\value{
Null
}
\description{
从vcfgztxt文件获取得到xlsx
}
\examples{
get_table("B3304.snp.VQSR.snpEff.vcf.gz.txt", paste("chr", c(1:10), sep = ""), mindp = 4)
}
