% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.gettxtfromvcf.R
\name{get_txt}
\alias{get_txt}
\title{Title}
\usage{
get_txt(sample_wt, sample_mut, vcffile)
}
\arguments{
\item{sample_wt}{野生型的样本名称}

\item{sample_mut}{突变体的样本名称}

\item{vcffile}{vcf文件的路径}
}
\value{
null
}
\description{
Title
}
\examples{
get_txt("B73QL", "B3304", "B3304.snp.VQSR.snpEff.vcf.gz")
}
