% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0.BSA.R
\name{pddepth}
\alias{pddepth}
\title{Title}
\usage{
pddepth(depthFT, mdepth)
}
\arguments{
\item{depthFT}{Whether to enable depth filtering (default "False"), use "True" to enable}

\item{mdepth}{specified sequencing depth threshold (default 0, specify non-zero value), effective when depthFT="True"}
}
\value{
depth threshold value
}
\description{
Title
}
\examples{
pddepth("True", 10)
}
