% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.run_vcf_processing.R
\name{run_vcf_processing}
\alias{run_vcf_processing}
\title{run_vcf_processing}
\usage{
run_vcf_processing(
  ref,
  refbwa,
  outdir,
  sample_mut,
  sample_wt,
  GFF3,
  mutfq1,
  mutfq2,
  wtfq1,
  wtfq2,
  depthFT = "False",
  mdepth = 0
)
}
\arguments{
\item{ref}{Path to reference genome FASTA file}

\item{refbwa}{Path to BWA-indexed reference genome (same as `ref`). If no index exists,
it will be created automatically. If pre-built index exists, specify the path: path/ref.fa}

\item{outdir}{Path to main output directory}

\item{sample_mut}{Mutant sample name (for result identification)}

\item{sample_wt}{Wild-type sample name (for result identification)}

\item{GFF3}{Path to genome annotation file (GFF3 format)}

\item{mutfq1}{Path to mutant sample FASTQ file (read 1)}

\item{mutfq2}{Path to mutant sample FASTQ file (read 2)}

\item{wtfq1}{Path to wild-type sample FASTQ file (read 1)}

\item{wtfq2}{Path to wild-type sample FASTQ file (read 2)}

\item{depthFT}{Whether to enable depth filtering (default "False"), use "True" to enable}

\item{mdepth}{specified sequencing depth threshold (default 0, specify non-zero value), effective when depthFT="True"}
}
\value{
No direct return value. Result files will be written to `outdir` directory, including:
}
\description{
run_vcf_processing
}
\examples{
run_vcf_processing(ref = "/path/to/reference.fa",refbwa = "/path/to/bwa_index_prefix",workdir = "/path/to/workdir",sample_mut = "MUT",sample_wt = "WT",mutfq1 = "/path/to/mut_1.fq.gz",mutfq2 = "/path/to/mut_2.fq.gz",wtfq1 = "/path/to/wt_1.fq.gz",wtfq2 = "/path/to/wt_2.fq.gz")
}
