% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.create_results.R
\name{create_results}
\alias{create_results}
\title{Title}
\usage{
create_results(
  outdir,
  sample_mut,
  sample_wt,
  depthFT = "False",
  mdepth = 0,
  GT_mut = "aa",
  mindp = 4
)
}
\arguments{
\item{GT_mut}{}
}
\description{
Title
}
