% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aa02.getxlsxfromtxt.R
\name{get_table}
\alias{get_table}
\title{Title get_table}
\usage{
get_table(
  file,
  chrlist,
  mindp = 4,
  minSNPindexmut = 0.3,
  maxSNPindexmut = 0.7,
  outdir,
  M
)
}
\arguments{
\item{file}{get_txt得到的txt文件路径}

\item{chrlist}{需要提取的染色体列表}

\item{mindp}{最低测序深度，默认4}

\item{minSNPindexmut}{最低SNPindex-M，默认0.3}

\item{maxSNPindexmut}{最高SNPindex-M，默认0.7}
}
\value{
Null
}
\description{
从vcfgztxt文件获取得到xlsx
}
\examples{
get_table("B3304.snp.VQSR.snpEff.vcf.gz.txt", paste("chr", c(1:10), sep = ""), mindp = 4)
}
