% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_software_path.R
\name{set_software_paths}
\alias{set_software_paths}
\title{Set or get software paths for BSA analysis}
\usage{
set_software_paths(
  bwa = NULL,
  fastp = NULL,
  samtools = NULL,
  picard = NULL,
  bcftools = NULL,
  gatk = NULL,
  snpEff = NULL,
  java = NULL,
  pandepth = NULL
)
}
\arguments{
\item{bwa}{Path to bwa executable}

\item{fastp}{Path to fastp executable}

\item{samtools}{Path to samtools executable}

\item{picard}{Path to picard jar file}

\item{bcftools}{Path to bcftools executable}

\item{gatk}{Path to gatk jar file}

\item{snpEff}{Path to snpEff directory}

\item{java}{Path to java executable}

\item{pandepth}{Path to pandepth executable}
}
\value{
A list of software paths (invisibly)
}
\description{
Set or get software paths for BSA analysis
}
